requirement.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
decorator==4.4.1
fire==0.2.1
Flask==1.1.1
idna==2.8
Inflector==3.0.1
itsdangerous==1.1.0
jellyfish==0.7.2
Jinja2==2.10.3
jmespath==0.9.4
langdetect==1.0.7
lxml==4.4.1
MarkupSafe==1.1.1
networkx==2.4
nltk==3.4.5
numpy==1.17.4
packaging==19.2
pandas==0.25.3
pdfminer==20191103
Pillow==6.2.1
pycryptodome==3.9.4
pyparsing==2.4.5
PyPDF2==1.26.0
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2019.3
regex==2019.11.1
requests==2.22.0
scipy==1.3.2
segtok==1.5.7
sip==5.0.0
six==1.13.0
termcolor==1.1.0
toml==0.10.0
Unidecode==1.1.1
urllib3==1.25.7
Werkzeug==0.16.0
xlrd==1.2.0
XlsxWriter==1.2.6